Reorganise Firebase classes
