+++ app/src/main/kotlin/com/jpaya/englishisfun/abbreviations/paging/AbbreviationsPageDataSource.kt
@@ -21,7 +21,7 @@ import androidx.annotation.VisibleForTesting.PRIVATE
 import androidx.lifecycle.MutableLiveData
 import androidx.paging.PageKeyedDataSource
 import com.jpaya.base.network.NetworkState
-import com.jpaya.englishisfun.abbreviations.firestore.FireStoreClient
+import com.jpaya.englishisfun.firestore.FireStoreClient
 import com.jpaya.englishisfun.abbreviations.model.AbbreviationItem
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
