--- base-android/src/main/kotlin/com/jpaya/base/firebase/FireStoreProperties.kt
@@ -1,39 +1,0 @@
-/*
- * Copyright 2020 Jose Maria Payá Castillo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.jpaya.base.firebase
-
-/**
- * Object that contains all FirebaseFireStore-related properties such as:
- *
- * - Collection names
- * - Document names
- * - Field names
- */
-class FireStoreProperties {
-
-    companion object {
-        private const val COLLECTION_NAME = "abbreviation"
-        private const val DOCUMENT_NAME = "list"
-        private const val LIST_FIELD = "abbreviations"
-    }
-
-    fun getAbbreviationCollectionName() = COLLECTION_NAME
-
-    fun getAbbreviationDocumentName() = DOCUMENT_NAME
-
-    fun getAbbreviationListField() = LIST_FIELD
-}
